# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Account.js.gripe Account API",
    "version": "2.0.0",
    "description": "JS.Gripe account-system v2 API. It uses UUIDv7 entity IDs and OAuth-style authorization code flow for third-party SSO."
  },
  "servers": [
    {
      "url": "https://gateway.js.gripe/api/v1/myaccount"
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service health"
          }
        }
      }
    },
    "/setup/status": {
      "get": {
        "summary": "Check whether the account system has an initial administrator",
        "responses": {
          "200": {
            "description": "Setup status"
          }
        }
      }
    },
    "/setup/init": {
      "post": {
        "summary": "Create the first system administrator",
        "responses": {
          "201": {
            "description": "Created administrator"
          },
          "409": {
            "description": "Already initialized"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Create an Account Center browser session",
        "responses": {
          "200": {
            "description": "Session bearer token for the Account Center UI"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register a member account",
        "responses": {
          "201": {
            "description": "Registered user and browser session"
          }
        }
      }
    },
    "/auth/authorize": {
      "post": {
        "summary": "Issue an authorization code for a registered client redirect",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Callback URL containing code and optional state"
          },
          "400": {
            "description": "Invalid redirect URI or scope"
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "summary": "Exchange an authorization code for a bearer token",
        "responses": {
          "200": {
            "description": "Bearer token, expiry, scope, and account user"
          },
          "400": {
            "description": "Invalid or expired authorization code"
          },
          "401": {
            "description": "Bad client credentials"
          }
        }
      }
    },
    "/clients/public": {
      "get": {
        "summary": "Get public API client metadata",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public client metadata"
          },
          "404": {
            "description": "Client not found"
          }
        }
      }
    },
    "/clients": {
      "get": {
        "summary": "List API clients",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Client list"
          }
        }
      },
      "post": {
        "summary": "Create an API client",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created client and one-time client secret"
          }
        }
      }
    },
    "/clients/{clientId}": {
      "delete": {
        "summary": "Delete an API client",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted client"
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Get the current authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/identity/resolve": {
      "get": {
        "summary": "Resolve a linked identity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved identity"
          },
          "404": {
            "description": "Identity not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}