# src/server.mjs
#!/usr/bin/env node
import fs from "node:fs/promises";
import http from "node:http";
import crypto from "node:crypto";
import path from "node:path";
import { fileURLToPath } from "node:url";
import { AccountStore } from "./store.mjs";

const rootDir = path.dirname(path.dirname(fileURLToPath(import.meta.url)));
const configPath = process.env.ACCOUNT_CONFIG_PATH || path.join(rootDir, "config", "account-system.env");
const dataPath = process.env.ACCOUNT_DB_PATH || path.join(rootDir, "data", "accounts.sqlite3");
const publicDir = process.env.ACCOUNT_PUBLIC_DIR || path.join(rootDir, "public");
const host = process.env.ACCOUNT_HOST || "127.0.0.1";
const port = Number(process.env.ACCOUNT_PORT || 9100);
const basePath = normalizeBasePath(process.env.ACCOUNT_BASE_PATH || "/api/v1/myaccount");
const allowedOrigin = process.env.ACCOUNT_ALLOWED_ORIGIN || "https://account.js.gripe";
const bootstrapTokenHash = process.env.ACCOUNT_BOOTSTRAP_TOKEN_HASH || "";

let store = new AccountStore(dataPath);
await store.load();
let bootstrapVerified = false;
let configuredDbPath = dataPath;

const server = http.createServer(async (req, res) => {
  try {
    const url = new URL(req.url || "/", `http://${req.headers.host || "localhost"}`);

    if (handleCors(req, res)) {
      return;
    }

    if (req.method === "GET" && (url.pathname === "/healthz" || url.pathname === `${basePath}/healthz`)) {
      sendJson(req, res, 200, { ok: true, service: "account-system", initialized: store.isInitialized() });
      return;
    }

    if (url.pathname === "/" || !url.pathname.startsWith(basePath)) {
      await serveStatic(req, res, url);
      return;
    }

    const routePath = url.pathname.slice(basePath.length) || "/";

    if (req.method === "GET" && routePath === "/setup/status") {
      sendJson(req, res, 200, {
        initialized: store.isInitialized(),
        bootstrap: bootstrapStatus()
      });
      return;
    }

    if (req.method === "POST" && routePath === "/setup/bootstrap/verify") {
      const body = await readJson(req);
      if (!bootstrapTokenHash) {
        sendJson(req, res, 503, { error: "bootstrap token is not configured; run npm run bootstrap:token on the server", code: "bootstrap_not_configured", detail: {} });
        return;
      }
      if (!verifyBootstrapToken(body.token)) {
        sendJson(req, res, 401, { error: "bootstrap token is invalid", code: "bad_bootstrap_token", detail: {} });
        return;
      }
      bootstrapVerified = true;
      sendJson(req, res, 200, { ok: true, bootstrap: bootstrapStatus() });
      return;
    }

    if (req.method === "POST" && routePath === "/setup/database/test") {
      requireBootstrap();
      const result = await testDatabasePath((await readJson(req)).dbPath);
      sendJson(req, res, 200, result);
      return;
    }

    if (req.method === "POST" && routePath === "/setup/database/apply") {
      requireBootstrap();
      const body = await readJson(req);
      const result = await testDatabasePath(body.dbPath);
      const nextStore = new AccountStore(result.dbPath);
      await nextStore.load();
      store = nextStore;
      configuredDbPath = result.dbPath;
      await writeServiceConfig({
        ACCOUNT_DB_PATH: configuredDbPath,
        ACCOUNT_BOOTSTRAP_TOKEN_HASH: bootstrapTokenHash
      });
      sendJson(req, res, 200, {
        ...result,
        bootstrap: bootstrapStatus(),
        initialized: store.isInitialized()
      });
      return;
    }

    if (req.method === "POST" && routePath === "/setup/init") {
      requireBootstrap();
      sendJson(req, res, 201, { user: await store.setupFirstAdmin(await readJson(req)) });
      return;
    }

    if (req.method === "POST" && routePath === "/auth/login") {
      if (!store.isInitialized()) {
        sendJson(req, res, 409, { error: "account system is not initialized" });
        return;
      }
      const body = await readJson(req);
      sendJson(req, res, 200, await store.login(body.email, body.password));
      return;
    }

    if (req.method === "POST" && routePath === "/auth/register") {
      if (!store.isInitialized()) {
        sendJson(req, res, 409, { error: "account system is not initialized" });
        return;
      }
      const body = await readJson(req);
      const user = await store.registerUser(body);
      const login = await store.login(user.email, body.password);
      sendJson(req, res, 201, login);
      return;
    }

    if (req.method === "GET" && routePath === "/clients/public") {
      const client = store.getClient(url.searchParams.get("client_id") || url.searchParams.get("clientId"));
      sendJson(req, res, client ? 200 : 404, client ? {
        client: {
          id: client.id,
          name: client.name,
          redirectUris: client.redirectUris,
          scopes: client.scopes,
          status: client.status
        }
      } : { error: "api client not found", code: "client_not_found", detail: {} });
      return;
    }

    if (req.method === "POST" && routePath === "/auth/token") {
      sendJson(req, res, 200, store.exchangeAuthorizationCode(await readJson(req)));
      return;
    }

    const session = requireSession(req);

    if (req.method === "POST" && routePath === "/auth/logout") {
      await store.logout(getBearerToken(req));
      sendJson(req, res, 200, { ok: true });
      return;
    }

    if (req.method === "POST" && routePath === "/auth/authorize") {
      sendJson(req, res, 200, store.authorizeClientSession({
        ...(await readJson(req)),
        expiresAt: session.expiresAt
      }, session.user));
      return;
    }

    if (req.method === "GET" && routePath === "/me") {
      sendJson(req, res, 200, { user: session.user, expiresAt: session.expiresAt });
      return;
    }

    if (req.method === "PATCH" && routePath === "/me") {
      sendJson(req, res, 200, { user: await store.updateOwnProfile(session.user.id, await readJson(req)) });
      return;
    }

    if (req.method === "POST" && routePath === "/me/password") {
      sendJson(req, res, 200, { user: await store.changeOwnPassword(session.user.id, await readJson(req)) });
      return;
    }

    if (req.method === "POST" && routePath === "/me/deactivate") {
      sendJson(req, res, 200, { user: await store.deactivateOwnAccount(session.user.id) });
      return;
    }

    if (req.method === "GET" && routePath === "/users") {
      requireCapability(session, "users");
      sendJson(req, res, 200, { users: store.listUsers() });
      return;
    }

    if (req.method === "POST" && routePath === "/users") {
      requireCapability(session, "users");
      sendJson(req, res, 201, { user: await store.createUser(await readJson(req), session.user.id) });
      return;
    }

    const userMatch = routePath.match(/^\/users\/([^/]+)$/);
    if (req.method === "GET" && userMatch) {
      requireCapability(session, "users");
      const user = store.getUser(userMatch[1]);
      sendJson(req, res, user ? 200 : 404, user ? { user } : { error: "user not found" });
      return;
    }

    if (req.method === "PATCH" && userMatch) {
      requireCapability(session, "users");
      sendJson(req, res, 200, { user: await store.updateUser(userMatch[1], await readJson(req), session.user.id) });
      return;
    }

    if (req.method === "DELETE" && userMatch) {
      requireCapability(session, "users");
      sendJson(req, res, 200, { user: await store.deleteUser(userMatch[1], session.user.id) });
      return;
    }

    const passwordMatch = routePath.match(/^\/users\/([^/]+)\/password$/);
    if (req.method === "POST" && passwordMatch) {
      requireCapability(session, "users");
      sendJson(req, res, 200, await store.resetPassword(passwordMatch[1], await readJson(req), session.user.id));
      return;
    }

    const linkMatch = routePath.match(/^\/users\/([^/]+)\/identities$/);
    if (req.method === "POST" && linkMatch) {
      requireCapability(session, "users");
      const identity = await store.linkIdentity(linkMatch[1], await readJson(req), session.user.id);
      sendJson(req, res, 201, { identity });
      return;
    }

    if (req.method === "DELETE" && routePath.startsWith("/identities/")) {
      requireCapability(session, "users");
      const identityId = routePath.split("/").at(-1);
      sendJson(req, res, 200, { identity: await store.unlinkIdentity(identityId, session.user.id) });
      return;
    }

    if (req.method === "GET" && routePath === "/identity/resolve") {
      const result = store.resolveIdentity(url.searchParams.get("provider"), url.searchParams.get("providerSubject"));
      if (result?.user?.status && result.user.status !== "active") {
        sendJson(req, res, 403, {
          error: result.user.disabledReason || "account is disabled",
          code: "account_disabled",
          detail: {
            userId: result.user.id,
            disabledReason: result.user.disabledReason || "",
            supportEmail: "helper@js.gripe"
          }
        });
        return;
      }
      sendJson(req, res, result ? 200 : 404, result || { error: "identity not found", code: "identity_not_found", detail: {} });
      return;
    }

    if (req.method === "GET" && routePath === "/clients") {
      requireCapability(session, "clients");
      const clients = store.listClients();
      sendJson(req, res, 200, { clients });
      return;
    }

    if (req.method === "POST" && routePath === "/clients") {
      requireCapability(session, "clients");
      const client = await store.createClient(await readJson(req), session.user.id);
      sendJson(req, res, 201, { client });
      return;
    }

    const apiMatch = routePath.match(/^\/clients\/([^/]+)$/);
    if (req.method === "DELETE" && apiMatch) {
      requireCapability(session, "clients");
      const client = await store.deleteClient(apiMatch[1], session.user.id);
      sendJson(req, res, 200, { client });
      return;
    }

    if (req.method === "POST" && routePath === "/clients/verify") {
      const body = await readJson(req);
      const client = store.verifyClient(body.clientId, body.clientSecret);
      sendJson(req, res, client ? 200 : 401, client ? { client } : { error: "bad api credentials", code: "bad_api_credentials", detail: {} });
      return;
    }

    if (req.method === "GET" && routePath === "/audit-logs") {
      requireCapability(session, "audit");
      sendJson(req, res, 200, { logs: store.listAuditLogs(url.searchParams.get("limit")) });
      return;
    }

    sendJson(req, res, 404, { error: "not found" });
  } catch (error) {
    sendJson(req, res, error.statusCode || 500, {
      error: error.message,
      code: error.code || "internal_error",
      detail: error.detail || {}
    });
  }
});

server.listen(port, host, () => {
  console.log(`account-system listening on http://${host}:${port}${basePath}`);
});

function requireSession(req) {
  const session = store.getSession(getBearerToken(req));
  if (!session) {
    const error = new Error("unauthorized");
    error.statusCode = 401;
    error.code = "unauthorized";
    throw error;
  }
  return session;
}

function requireCapability(session, capability) {
  if (!session.user.capabilities?.[capability]) {
    const error = new Error(`${capability} permission required`);
    error.statusCode = 403;
    error.code = "permission_denied";
    throw error;
  }
}

function requireBootstrap() {
  if (store.isInitialized()) {
    return;
  }
  if (!bootstrapVerified) {
    const error = new Error("bootstrap token verification is required");
    error.statusCode = 403;
    error.code = "bootstrap_required";
    throw error;
  }
}

function verifyBootstrapToken(token) {
  if (!bootstrapTokenHash) {
    return false;
  }
  const actual = Buffer.from(hashSecret(String(token || "")));
  const expected = Buffer.from(bootstrapTokenHash);
  return actual.length === expected.length && crypto.timingSafeEqual(actual, expected);
}

function bootstrapStatus() {
  const initialized = store.isInitialized();
  return {
    required: !initialized,
    configured: Boolean(bootstrapTokenHash),
    verified: initialized || bootstrapVerified,
    dbPath: configuredDbPath
  };
}

async function testDatabasePath(inputPath) {
  const dbPath = normalizeDbPath(inputPath);
  await fs.mkdir(path.dirname(dbPath), { recursive: true });
  const probe = new AccountStore(dbPath);
  await probe.load();
  return {
    ok: true,
    dbPath,
    initialized: probe.isInitialized()
  };
}

function normalizeDbPath(inputPath) {
  const raw = String(inputPath || configuredDbPath || dataPath).trim();
  if (!raw) {
    const error = new Error("database path is required");
    error.statusCode = 400;
    error.code = "db_path_required";
    throw error;
  }
  return path.isAbsolute(raw) ? path.normalize(raw) : path.resolve(rootDir, raw);
}

async function writeServiceConfig(values) {
  await fs.mkdir(path.dirname(configPath), { recursive: true });
  let existing = "";
  try {
    existing = await fs.readFile(configPath, "utf8");
  } catch (error) {
    if (error.code !== "ENOENT") throw error;
  }

  const lines = existing
    .split(/\r?\n/)
    .filter((line) => line.trim() && !Object.keys(values).some((key) => line.startsWith(`${key}=`)));
  for (const [key, value] of Object.entries(values)) {
    if (value) lines.push(`${key}=${quoteEnv(value)}`);
  }
  await fs.writeFile(configPath, `${lines.join("\n")}\n`, { encoding: "utf8", mode: 0o600 });
}

function quoteEnv(value) {
  return `'${String(value).replaceAll("'", "'\\''")}'`;
}

function hashSecret(secret) {
  return crypto.createHash("sha256").update(secret).digest("hex");
}

function getBearerToken(req) {
  const value = req.headers.authorization || "";
  return value.startsWith("Bearer ") ? value.slice(7).trim() : "";
}

function readJson(req) {
  return new Promise((resolve, reject) => {
    const chunks = [];
    let size = 0;

    req.on("data", (chunk) => {
      size += chunk.length;
      if (size > 1024 * 1024) {
        const error = new Error("payload too large");
        error.statusCode = 413;
        reject(error);
        req.destroy();
        return;
      }
      chunks.push(chunk);
    });

    req.on("end", () => {
      try {
        const raw = Buffer.concat(chunks).toString("utf8");
        resolve(raw ? JSON.parse(raw) : {});
      } catch {
        const error = new Error("invalid json");
        error.statusCode = 400;
        reject(error);
      }
    });
    req.on("error", reject);
  });
}

async function serveStatic(req, res, url) {
  if (req.method !== "GET" && req.method !== "HEAD") {
    sendJson(req, res, 405, { error: "method not allowed" });
    return;
  }

  const cleanPath = decodeURIComponent(url.pathname).replace(/^\/+/, "");
  const routeFile = {
    "": "login.html",
    login: "login.html",
    dashboard: "dashboard.html"
  }[cleanPath] || cleanPath;
  const requestedPath = path.normalize(routeFile || "login.html");
  if (requestedPath.startsWith("..")) {
    sendJson(req, res, 403, { error: "forbidden" });
    return;
  }

  let filePath = path.join(publicDir, requestedPath);
  try {
    const stat = await fs.stat(filePath);
    if (stat.isDirectory()) {
      filePath = path.join(filePath, "index.html");
    }
  } catch {
    filePath = path.join(publicDir, "index.html");
  }

  const body = await fs.readFile(filePath);
  const type = contentType(filePath);
  res.writeHead(200, {
    "content-type": type,
    "cache-control": cacheControlForStatic(filePath, type),
    "content-length": body.length
  });
  if (req.method === "HEAD") {
    res.end();
    return;
  }
  res.end(body);
}

function sendJson(req, res, statusCode, body) {
  const payload = JSON.stringify(body);
  res.writeHead(statusCode, {
    "content-type": "application/json; charset=utf-8",
    "content-length": Buffer.byteLength(payload),
    "cache-control": "no-store",
    ...corsHeaders(req)
  });
  res.end(payload);
}

function handleCors(req, res) {
  if (req.method !== "OPTIONS") {
    return false;
  }
  res.writeHead(204, {
    ...corsHeaders(req),
    "access-control-allow-methods": "GET, POST, PATCH, DELETE, OPTIONS",
    "access-control-allow-headers": "Authorization, Content-Type",
    "access-control-max-age": "86400"
  });
  res.end();
  return true;
}

function corsHeaders(req) {
  const origin = req.headers.origin;
  if (origin !== allowedOrigin) {
    return {};
  }
  return {
    "access-control-allow-origin": allowedOrigin,
    "vary": "Origin"
  };
}

function contentType(filePath) {
  const name = path.basename(filePath);
  if (name === "api-catalog") return "application/linkset+json; charset=utf-8";
  if (filePath.includes(`${path.sep}.well-known${path.sep}`) && !path.extname(filePath)) return "application/json; charset=utf-8";
  if (filePath.endsWith(".css")) return "text/css; charset=utf-8";
  if (filePath.endsWith(".js")) return "text/javascript; charset=utf-8";
  if (filePath.endsWith(".json")) return "application/json; charset=utf-8";
  if (filePath.endsWith(".md")) return "text/markdown; charset=utf-8";
  if (filePath.endsWith(".txt")) return "text/plain; charset=utf-8";
  if (filePath.endsWith(".xml")) return "application/xml; charset=utf-8";
  if (filePath.endsWith(".svg")) return "image/svg+xml";
  if (filePath.endsWith(".png")) return "image/png";
  if (filePath.endsWith(".jpg") || filePath.endsWith(".jpeg")) return "image/jpeg";
  if (filePath.endsWith(".webp")) return "image/webp";
  if (filePath.endsWith(".ico")) return "image/x-icon";
  return "text/html; charset=utf-8";
}

function cacheControlForStatic(filePath, type) {
  const name = path.basename(filePath);
  if (filePath.endsWith("index.html") || name === "login.html" || name === "dashboard.html" || type.startsWith("text/javascript")) {
    return "no-store";
  }
  return "public, max-age=3600";
}

function normalizeBasePath(value) {
  const trimmed = String(value || "").trim();
  if (!trimmed || trimmed === "/") {
    return "";
  }
  return trimmed.startsWith("/") ? trimmed.replace(/\/+$/, "") : `/${trimmed.replace(/\/+$/, "")}`;
}